Generic non-recursive multiset combinations
I have ported the non-recursive method for generating multiset
combinations to a pair of generic functions which can be used to
generate such combinations out of any vector of objects.  Broke the
functions into multichoose.hpp, which is suitable for import into other
projects.  One version (multiset_ptr) returns results as a vector of
vector of pointers to the original objects.  This may be useful in the
case of large objects; at present I have not tested.