Correct index advancement arithmetic
By decrementing j in a do { ... } while loop we avoid a bug whereby
comparing potentially out-of-bounds (and unequal) memory locations can
trigger an infinite loop.

Previously, the algorithm never decremented j in some cases; and in
every case the first comparison in the while loop to do so was out of
bounds!

It seems that out-of-bounds locations are often initialized to the same
value.  This case would have allowed the algorithm to continue as
normal.  Otherwise I don't understand any reason why this bug would
manifest itself irregularly.  The issue did not appear at all in tests
against multichoose.cpp until I switched it to use a vector<string>
instead of a vector<string*> as the input to multichoose.  And after
than it only appeared sporadically; and most regularly when I integrated
multichoose into a larger application.